cleans up settings page
