--- src/app/[lang]/(protected)/settings/templates/page.tsx
@@ -1,95 +1,0 @@
-// /settings/templates/page.tsx
-
-import { Suspense } from "react";
-import { TemplatesClient } from "@/components/settings/waba-templates";
-import { getServerSession } from "next-auth";
-import { authOptions } from "@/lib/auth";
-import PageTitle from "@/components/ui/page-title";
-import { isValidLanguage } from "@/lib/i18n";
-import {
-  WabaTemplateFilters,
-  WabaTemplateRepository,
-} from "@/data/repositories/waba-template";
-import { getUserByIdAction } from "@/actions/user.actions";
-import { Prisma } from "@/lib/prisma/generated";
-
-function TemplatesSkeleton() {
-  return (
-    <div className="space-y-4">
-      {[1, 2, 3].map((i) => (
-        <div key={i} className="border rounded-lg p-6 animate-pulse">
-          <div className="h-6 bg-foreground rounded w-1/3 mb-4"></div>
-          <div className="h-4 bg-foreground rounded w-2/3 mb-2"></div>
-          <div className="h-4 bg-foreground rounded w-1/2"></div>
-        </div>
-      ))}
-    </div>
-  );
-}
-
-const translations = {
-  en: {
-    title: "WhatsApp Templates",
-    subtitle: "Manage your WhatsApp Business API message templates",
-  },
-  sw: {
-    title: "Violezo vya WhatsApp",
-    subtitle: "Dhibiti violezo vya ujumbe wa WhatsApp Business API",
-  },
-};
-
-export default async function TemplatesPage({
-  params,
-}: {
-  params: Promise<{ lang: string }>;
-}) {
-  const { lang } = await params;
-  const session = await getServerSession(authOptions);
-  if (!session) return;
-  if (!isValidLanguage(lang)) return;
-
-  const t = translations[lang];
-
-  const templates = await WabaTemplateRepository.findByUserId(session.user.id);
-
-  // const wabaAccount = (await getUserById(session.user.id))?.waba || null;
-
-  const resUserWithWaba = await getUserByIdAction(session.user.id, {
-    waba: {
-      include: {
-        phoneNumbers: true,
-        templates: true,
-      },
-    },
-  });
-
-  let userWithWaba: Prisma.UserGetPayload<{
-    include: {
-      waba: {
-        include: {
-          phoneNumbers: true;
-          templates: true;
-        };
-      };
-    };
-  }> | null = null;
-  let wabaAccount: Prisma.WabaAccountGetPayload<{
-    include: {
-      phoneNumbers: true;
-      templates: true;
-    };
-  }> | null = null;
-
-  if (resUserWithWaba.ok) userWithWaba = resUserWithWaba.data;
-  if (userWithWaba) wabaAccount = userWithWaba.waba;
-
-  return (
-    <div className="max-w-7xl mx-auto space-y-6 py-6">
-      <PageTitle title={t.title} subtitle={t.subtitle} />
-
-      <Suspense fallback={<TemplatesSkeleton />}>
-        <TemplatesClient wabaAccount={wabaAccount} />
-      </Suspense>
-    </div>
-  );
-}
