+++ src/app/[lang]/(public)/pricing/page.tsx
@@ -14,13 +14,13 @@ export default async function Home({
   return (
     <main>
       <section id="pricing" className="relative py-12 md:py-20">
-        <div className="relative max-w-screen-xl mx-auto px-4 sm:px-6">
+        <div className="relative max-w-7xl mx-auto px-4 sm:px-6">
           {/* Section header */}
           <div className="max-w-3xl mx-auto text-center pb-6">
             <h2 className="text-3xl md:text-4xl font-bold mb-4">{t.title}</h2>
             <p className="text-lg text-muted-foreground">{t.subtitle}</p>
           </div>
-          <PricingComponent t={t} />
+          <PricingComponent />
         </div>
       </section>
     </main>
