+++ src/components/settings/settings-client/index.tsx
@@ -2,28 +2,9 @@
 
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { Separator } from "@/components/ui/separator";
-import {
   User as UserIcon,
-  Phone,
   CreditCard,
-  Bell,
   Shield,
-  Building2,
-  Mail,
-  Calendar,
-  CheckCircle2,
-  XCircle,
-  AlertCircle,
   MessageSquareIcon,
 } from "lucide-react";
 import { WhatsAppLogo } from "@/components/ui/svg";
